Make Ident a functor - Huge *breaking* refactor
I needed to make Ident a functor so that I could errase
all annotations. This, however, had some side-effects that
made refactor the AST in a breaking way. But now it is possible
to derive 'functor' for all objects in the AST.

The main changes are in `src/Language/C/Syntax/AST.hs`